+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/command/application/controller/SeatController.java
@@ -24,12 +24,12 @@ public class SeatController {
         좌석 접수
      */
     @PostMapping("/{concertId}/seats/{seatId}/reception")
-    public ResponseEntity<?> seatReception(@PathVariable("concertId") int concertId,
-                                           @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> seatReception(@PathVariable("concertId") Long concertId,
+                                           @PathVariable("seatId") Long seatId) {
 
         log.info("seatReception_request: concertId={}, seatId={}", concertId, seatId);
 
-        SeatResponseDTO.seatReceptionDTO responseDTO = seatService.seatReception((long) concertId, (long) seatId, getCurrentMemberId());
+        SeatResponseDTO.seatReceptionDTO responseDTO = seatService.seatReception(concertId, seatId, getCurrentMemberId());
 
         log.info("seatReception_response: {}", responseDTO);
 
@@ -40,12 +40,12 @@ public class SeatController {
         좌석 접수 취소
      */
     @DeleteMapping("/{concertId}/seats/{seatId}")
-    public ResponseEntity<?> cancelReceptionSeat(@PathVariable("concertId") int concertId,
-                                                 @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> cancelReceptionSeat(@PathVariable("concertId") Long concertId,
+                                                 @PathVariable("seatId") Long seatId) {
 
         log.info("cancelReceptionSeat_request: concertId={}, seatId={}", concertId, seatId);
 
-        SeatResponseDTO.cancelReceptionSeatDTO responseDTO = seatService.cancelReceptionSeat((long) concertId, (long) seatId, getCurrentMemberId());
+        SeatResponseDTO.cancelReceptionSeatDTO responseDTO = seatService.cancelReceptionSeat(concertId, seatId, getCurrentMemberId());
 
         log.info("cancelReceptionSeat_response: {}", responseDTO);
 
@@ -56,12 +56,12 @@ public class SeatController {
         추첨 시작 알림
      */
     @PostMapping("/{concertId}/seats/{seatId}/drawing")
-    public ResponseEntity<?> drawingNotification(@PathVariable("concertId") int concertId,
-                                                 @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> drawingNotification(@PathVariable("concertId") Long concertId,
+                                                 @PathVariable("seatId") Long seatId) {
 
         log.info("drawingNotification_request: concertId={}, seatId={}", concertId, seatId);
 
-        SeatResponseDTO.createDrawingNotificationDTO responseDTO = seatService.createDrawingNotification((long) concertId, (long) seatId);
+        SeatResponseDTO.createDrawingNotificationDTO responseDTO = seatService.createDrawingNotification(concertId, seatId);
 
         log.info("drawingNotification_response: {}", responseDTO);
 
@@ -72,12 +72,12 @@ public class SeatController {
         좌석 추첨 결과 반영
      */
     @PostMapping("/{concertId}/seats/{seatId}/result")
-    public ResponseEntity<?> processDrawResult(@PathVariable("concertId") int concertId,
-                                               @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> processDrawResult(@PathVariable("concertId") Long concertId,
+                                               @PathVariable("seatId") Long seatId) {
 
         log.info("createDrawResult_request: concertId={}, seatId={}", concertId, seatId);
 
-        seatService.processDrawResult((long) concertId, (long) seatId, getCurrentMemberId());
+        seatService.processDrawResult(concertId, seatId, getCurrentMemberId());
 
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
@@ -86,9 +86,9 @@ public class SeatController {
         좌석 추첨 결과 - 치트
      */
     @PostMapping("/{concertId}/draw-cheat")
-    public ResponseEntity<?> cheatDrawResult(@PathVariable("concertId") int concertId) {
+    public ResponseEntity<?> cheatDrawResult(@PathVariable("concertId") Long concertId) {
 
-        seatService.cheatDrawResult((long) concertId, getCurrentMemberId());
+        seatService.cheatDrawResult(concertId, getCurrentMemberId());
 
         return ResponseEntity.ok().body(ApiUtils.success(null));
     }
@@ -97,12 +97,12 @@ public class SeatController {
         좌석 결제
      */
     @PostMapping("/{concertId}/seats/{seatId}/payment")
-    public ResponseEntity<?> reserveSeat(@PathVariable("concertId") int concertId,
-                                         @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> reserveSeat(@PathVariable("concertId") Long concertId,
+                                         @PathVariable("seatId") Long seatId) {
 
         log.info("reserveSeat_request: concertId={}, seatId={}", concertId, seatId);
 
-        SeatResponseDTO.reserveSeatDTO responseDTO = seatService.reserveSeat((long) concertId, (long) seatId, getCurrentMemberId());
+        SeatResponseDTO.reserveSeatDTO responseDTO = seatService.reserveSeat(concertId, seatId, getCurrentMemberId());
 
         log.info("reserveSeat_response: {}", responseDTO);
 
@@ -113,9 +113,9 @@ public class SeatController {
         좌석 결제 - 치트
      */
     @PostMapping("/{concertId}/payment-cheat")
-    public ResponseEntity<?> cheatReserveSeat(@PathVariable("concertId") int concertId) {
+    public ResponseEntity<?> cheatReserveSeat(@PathVariable("concertId") Long concertId) {
 
-        SeatResponseDTO.reserveSeatDTO responseDTO = seatService.cheatReserveSeat((long) concertId, getCurrentMemberId());
+        SeatResponseDTO.reserveSeatDTO responseDTO = seatService.cheatReserveSeat(concertId, getCurrentMemberId());
 
         return ResponseEntity.ok().body(ApiUtils.success(responseDTO));
     }
