+++ src/main/java/ticketaka/mtvs3_final_backend/ticketing/seat/query/controller/SeatQueryController.java
@@ -27,12 +27,12 @@ public class SeatQueryController {
         좌석 조회
      */
     @GetMapping("/{concertId}/seats/{seatId}")
-    public ResponseEntity<?> getConcertSeat(@PathVariable("concertId") int concertId,
-                                            @PathVariable("seatId") int seatId) {
+    public ResponseEntity<?> getConcertSeat(@PathVariable("concertId") Long concertId,
+                                            @PathVariable("seatId") Long seatId) {
 
         log.info("GetConcertSeat_request: concertId={}, seatId={}", concertId, seatId);
 
-        SeatQueryResponseDTO.getSeatInfoDTO responseDTO = seatQueryService.getConcertSeat((long) concertId, (long) seatId);
+        SeatQueryResponseDTO.getSeatInfoDTO responseDTO = seatQueryService.getConcertSeat(concertId, seatId);
 
         log.info("GetConcertSeat_response: {}", responseDTO);
 
@@ -43,7 +43,7 @@ public class SeatQueryController {
         현재 회원이 접수한 좌석 조회
      */
     @GetMapping("{concertId}/receptions")
-    public ResponseEntity<?> getMyConcertReceptions(@PathVariable("concertId") long concertId) {
+    public ResponseEntity<?> getMyConcertReceptions(@PathVariable("concertId") Long concertId) {
 
         log.info("GetMyConcertReceptions_request: concertId={}", concertId);
